feat: topup page separate client component
